SWPROT-8953: Disable DEV_GUI as not reproducible
I suspect a floating dep, it may be fixed in later when stable,
ideally this should be located elsewhere to avoid breaking.

The fix may be trivial, with npm audit:

    +        "typescript": "4.7.4",
    +        "react-router-dom": "5.3.3",

But for history purposes we wont touch files that are not z-wave
in this repo.

Origin: https://github.com/SiliconLabs/UnifySDK/pull/50
Forwarded: https://github.com/SiliconLabsSoftware/unify-z-wave/pull/3
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>