UIC-8953: rust: Allow target triple overide in CMakeCargo too
This change is helpful to trace/debug rust related job.
This was useful to build zpc natively on arm64 (aarch64-unknown-linux-gnu).

The earlier change is not enough.

What has been observed on native arm64 build:

    .../1.71.0-aarch64-unknown-linux-gnu/bin/cargo build \
   --manifest-path /.../applications/zpc/components/zpc_rust/Cargo.toml \
   --locked \
   --target x86_64-unknown-linux-gnu

While it was passed a cmake define

Note to self, for maintenance sake,
it would make sense to contribute those macros outside,
the project ideally in rust or cmake community.

It was made for core v7 but also applies to zpc v6

Relate-to: https://doc.rust-lang.org/nightly/rustc/platform-support.html
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>