SWPROT-8953: applications/home_assistant: Removed as not part of z-wave-protocol-controller
This component is maintained at UnifySDK (v1.7+ aka core),

This change deduplicate code, to ensure a single source of truth
and enable a more sustainable developement model
without overhead of monorepo.

This change may be not atomic,
but has been done in a way one component could be reverted
if it justify a downstream (double) maintainance or experimental works.

For the record this component was imported from:
https://github.com/SiliconLabs/UnifySDK/releases/tag/ver_1.6.0

Docs should be regenerated from sources, but not stored along sources.

More cleanup might be desirable for 1.7.

Bug-SiliconLabs: SWPROT-8953
Origin: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/pull/17
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>