SWPROT-8953: doc/getting_started_multiprotocol_cpc.md: Removed as not part of z-wave-protocol-controller
This component is maintained at UnifySDK (v1.7+ aka core),

This change deduplicate code, to ensure a single source of truth
and enable a more sustainable developement model
without overhead of monorepo.

This change may be not atomic,
but has been done in a way one component could be reverted
if it justify a downstream (double) maintainance or experimental works.

For the record this component was imported from:
https://github.com/SiliconLabs/UnifySDK/releases/tag/ver_1.6.0

If this component should be re-introduced it would be preferable to
consider the code where it is maintained eg:

https://github.com/SiliconLabs/UnifySDK/releases/tag/ver_1.7.0 (or later)

Bug-SiliconLabs: SWPROT-8953
Origin: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/pull/36
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>