build: p/UnifySDK: Add patch to remove some assets #77
For the record, I have enabled github action in UnifySDK repo,
It is actually runing in forks eg:

https://github.com/rzr/UnifySDK/tree/review/main

https://github.com/rzr/UnifySDK/actions/runs/13765559449/job/38491058218

Tests are passing:

But it fail later on packaging:

I suspect the cause it all the assets in tree, this has to be cleaned up.

This change prevent this problem, further cleaning may be considered.

Origin: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/pull/47
Relate-to: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/issues/77
Relate-to: https://github.com/SiliconLabsSoftware/z-wave-engine-application-layer/issues/10
Relate-to: https://github.com/SiliconLabsSoftware/z-wave-engine-application-layer/issues/11
Relate-to: https://github.com/SiliconLabsSoftware/z-wave-engine-application-layer/issues/7
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>