fix(tests): Add HOMEID support for cli and improve debug
Wait for broker to be ready before starting play

I noticed that some test session where failing to start,
if failed if broker was not ready, now it retries it.

Improving debug of z-wave-stack-binaries-test.sh

This is a minor change to test CI, I should expect this commit to be tested.

Origin: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/pull/92
Relate-to: https://github.com/SiliconLabsSoftware/z-wave-engine-application-layer/issues/23
Relate-to: https://github.com/Z-Wave-Alliance/z-wave-stack/issues/732
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>